--- todos-gateway.js
@@ -1,7 +1,0 @@
-module.exports = function todosGateway(collection) {
-  return {
-    async find(where) {
-      return collection.find(where).toArray()
-    }
-  }
-}
